The "riverdensity" Class

A class that holds density information computed from point data along a river network.

Details

Created by makeriverdensity from point data and a river network. Contains all information for plotting in plot.riverdensity.

Elements

densities:

Object of class "list". Each list element corresponds to a unique value of survey. Each element is itself of class "list", with each element corresponding to a segment from the associated river network. Each element is a vector of class "numeric", with values equal to the scaled densities calculated at the river network vertices stored in $densverts of the associated river network segment.

endptverts:

List of vectors of class "numeric". Each list element is a vector of the vertices of the endpoints of the subsegments considered for density calculation. Each list element corresponds to a river segment from the associated river network.

densverts:

List of vectors of class "numeric". Each element is a vector of the vertices of the points of the subsegments considered for density calculation, that were used for density calculation. Each list element corresponds to a river segment from the associated river network.

pointsegs:

Vector of class "numeric". Defined as the segment numbers of the point data used for density calculation.

pointverts:

Vector of class "numeric". Defined as the vertex numbers of the point data used for density calculation.

survey:

Vector of class "numeric" or class "character". Defined as the survey identifiers associated with the point data used for density calculation.

rivers:

Object of class "rivernetwork" ; see rivernetwork-class

.