Store Vertices in Ascending Sequence

Rearranges the vertices of a river network object so that vertices are stored sequentially moving up river for all segments (coordinates [1,] are the bottom of each segment).

sequenceverts(rivers)

Arguments

rivers
The river network object to use

Value

A new river network object (see rivernetwork)

Note

Even without calling sequenceverts, the vertices will be stored sequentially - either moving up river or down for a given segment. What sequenceverts() adds is a standardized direction.

Currently, no function in package 'riverdist' requires the vertices to be stored sequentially.

See also

line2network

Examples

data(Gulk)
Gulk <- setmouth(seg=1, vert=1, rivers=Gulk)
str(Gulk)

#> List of 14
#>  $ sp         :Formal class 'SpatialLinesDataFrame' [package "sp"] with 4 slots
#>   .. ..@ data       :'data.frame':	1 obs. of  1 variable:
#>   .. .. ..$ Id: int 0
#>   .. ..@ lines      :List of 1
#>   .. .. ..$ :Formal class 'Lines' [package "sp"] with 2 slots
#>   .. .. .. .. ..@ Lines:List of 14
#>   .. .. .. .. .. ..$ :Formal class 'Line' [package "sp"] with 1 slot
#>   .. .. .. .. .. .. .. ..@ coords: num [1:812, 1:2] 894793 894701 894598 894570 894516 ...
#>   .. .. .. .. .. ..$ :Formal class 'Line' [package "sp"] with 1 slot
#>   .. .. .. .. .. .. .. ..@ coords: num [1:33, 1:2] 864906 864892 864901 864899 864887 ...
#>   .. .. .. .. .. ..$ :Formal class 'Line' [package "sp"] with 1 slot
#>   .. .. .. .. .. .. .. ..@ coords: num [1:327, 1:2] 878280 878142 878076 878029 877963 ...
#>   .. .. .. .. .. ..$ :Formal class 'Line' [package "sp"] with 1 slot
#>   .. .. .. .. .. .. .. ..@ coords: num [1:674, 1:2] 878280 878281 878285 878255 878247 ...
#>   .. .. .. .. .. ..$ :Formal class 'Line' [package "sp"] with 1 slot
#>   .. .. .. .. .. .. .. ..@ coords: num [1:43, 1:2] 865525 865507 865504 865505 865492 ...
#>   .. .. .. .. .. ..$ :Formal class 'Line' [package "sp"] with 1 slot
#>   .. .. .. .. .. .. .. ..@ coords: num [1:488, 1:2] 865525 865432 865363 865266 865205 ...
#>   .. .. .. .. .. ..$ :Formal class 'Line' [package "sp"] with 1 slot
#>   .. .. .. .. .. .. .. ..@ coords: num [1:61, 1:2] 851642 851558 851522 851454 851414 ...
#>   .. .. .. .. .. ..$ :Formal class 'Line' [package "sp"] with 1 slot
#>   .. .. .. .. .. .. .. ..@ coords: num [1:543, 1:2] 851642 851635 851593 851511 851462 ...
#>   .. .. .. .. .. ..$ :Formal class 'Line' [package "sp"] with 1 slot
#>   .. .. .. .. .. .. .. ..@ coords: num [1:786, 1:2] 873422 873406 873385 873391 873405 ...
#>   .. .. .. .. .. ..$ :Formal class 'Line' [package "sp"] with 1 slot
#>   .. .. .. .. .. .. .. ..@ coords: num [1:41, 1:2] 873422 873403 873280 873263 873239 ...
#>   .. .. .. .. .. ..$ :Formal class 'Line' [package "sp"] with 1 slot
#>   .. .. .. .. .. .. .. ..@ coords: num [1:454, 1:2] 872726 872587 872560 872436 872387 ...
#>   .. .. .. .. .. ..$ :Formal class 'Line' [package "sp"] with 1 slot
#>   .. .. .. .. .. .. .. ..@ coords: num [1:484, 1:2] 872726 872710 872706 872668 872617 ...
#>   .. .. .. .. .. ..$ :Formal class 'Line' [package "sp"] with 1 slot
#>   .. .. .. .. .. .. .. ..@ coords: num [1:869, 1:2] 860552 860458 860428 860385 860210 ...
#>   .. .. .. .. .. ..$ :Formal class 'Line' [package "sp"] with 1 slot
#>   .. .. .. .. .. .. .. ..@ coords: num [1:361, 1:2] 860552 860545 860490 860488 860504 ...
#>   .. .. .. .. ..@ ID   : chr "0"
#>   .. ..@ bbox       : num [1:2, 1:2] 839774 6922087 895655 7010276
#>   .. .. ..- attr(*, "dimnames")=List of 2
#>   .. .. .. ..$ : chr [1:2] "x" "y"
#>   .. .. .. ..$ : chr [1:2] "min" "max"
#>   .. ..@ proj4string:Formal class 'CRS' [package "sp"] with 1 slot
#>   .. .. .. ..@ projargs: chr "+proj=utm +zone=5 +datum=NAD83 +units=m +no_defs +ellps=GRS80 +towgs84=0,0,0"
#>  $ lines      :List of 14
#>   ..$ : num [1:812, 1:2] 894793 894701 894598 894570 894516 ...
#>   ..$ : num [1:33, 1:2] 864906 864892 864901 864899 864887 ...
#>   ..$ : num [1:327, 1:2] 878280 878142 878076 878029 877963 ...
#>   ..$ : num [1:674, 1:2] 878280 878281 878285 878255 878247 ...
#>   ..$ : num [1:43, 1:2] 865525 865507 865504 865505 865492 ...
#>   ..$ : num [1:488, 1:2] 865525 865432 865363 865266 865205 ...
#>   ..$ : num [1:61, 1:2] 851642 851558 851522 851454 851414 ...
#>   ..$ : num [1:543, 1:2] 851642 851635 851593 851511 851462 ...
#>   ..$ : num [1:786, 1:2] 873422 873406 873385 873391 873405 ...
#>   ..$ : num [1:41, 1:2] 873422 873403 873280 873263 873239 ...
#>   ..$ : num [1:454, 1:2] 872726 872587 872560 872436 872387 ...
#>   ..$ : num [1:484, 1:2] 872726 872710 872706 872668 872617 ...
#>   ..$ : num [1:869, 1:2] 860552 860458 860428 860385 860210 ...
#>   ..$ : num [1:361, 1:2] 860552 860545 860490 860488 860504 ...
#>  $ connections: num [1:14, 1:14] NA NA 2 2 NA NA NA NA NA NA ...
#>  $ lengths    : num [1:14] 76631 1066 27718 52881 1486 ...
#>  $ names      : logi [1:14] NA NA NA NA NA NA ...
#>  $ mouth      :List of 2
#>   ..$ mouth.seg : num 1
#>   ..$ mouth.vert: num 1
#>  $ sequenced  : logi FALSE
#>  $ tolerance  : num 100
#>  $ units      : chr "m"
#>  $ lineID     :'data.frame':	14 obs. of  3 variables:
#>   ..$ rivID  : int [1:14] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ sp_line: num [1:14] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ sp_seg : int [1:14] 1 2 3 4 5 6 7 8 9 10 ...
#>  $ braided    : logi FALSE
#>  $ segroutes  :List of 14
#>   ..$ : int 1
#>   ..$ : int [1:4] 1 3 5 2
#>   ..$ : int [1:2] 1 3
#>   ..$ : int [1:2] 1 4
#>   ..$ : int [1:3] 1 3 5
#>   ..$ : int [1:3] 1 3 6
#>   ..$ : int [1:4] 1 3 6 7
#>   ..$ : int [1:4] 1 3 6 8
#>   ..$ : int [1:3] 1 4 9
#>   ..$ : int [1:3] 1 4 10
#>   ..$ : int [1:4] 1 4 10 11
#>   ..$ : int [1:4] 1 4 10 12
#>   ..$ : int [1:5] 1 4 10 11 13
#>   ..$ : int [1:5] 1 4 10 11 14
#>  $ cumuldist  :List of 14
#>   ..$ : num [1:812] 0 93.2 211.3 249.3 303 ...
#>   ..$ : num [1:33] 0 21.7 71.8 83.5 100.6 ...
#>   ..$ : num [1:327] 0 141 223 271 338 ...
#>   ..$ : num [1:674] 0 22.4 99.2 199.6 276.7 ...
#>   ..$ : num [1:43] 0 91.6 106.9 173.3 202.3 ...
#>   ..$ : num [1:488] 0 113 185 283 357 ...
#>   ..$ : num [1:61] 0 115 159 235 309 ...
#>   ..$ : num [1:543] 0 45.7 130.4 225.8 274.8 ...
#>   ..$ : num [1:786] 0 19.3 52.3 92.7 123.8 ...
#>   ..$ : num [1:41] 0 33.7 197.4 235.2 348.2 ...
#>   ..$ : num [1:454] 0 140 172 300 349 ...
#>   ..$ : num [1:484] 0 31.4 66.4 116 169.4 ...
#>   ..$ : num [1:869] 0 94.7 129.8 172.4 360.5 ...
#>   ..$ : num [1:361] 0 38.7 118.6 146.7 171.6 ...
#>  $ distlookup :List of 3
#>   ..$ middist : num [1:14, 1:14] 0 29204 0 0 27718 ...
#>   ..$ starttop: logi [1:14, 1:14] NA TRUE TRUE TRUE TRUE TRUE ...
#>   ..$ endtop  : logi [1:14, 1:14] NA FALSE FALSE FALSE FALSE FALSE ...
#>  - attr(*, "class")= chr "rivernetwork"

Gulk.dir <- sequenceverts(rivers=Gulk)
str(Gulk.dir)

#> List of 14
#>  $ sp         :Formal class 'SpatialLinesDataFrame' [package "sp"] with 4 slots
#>   .. ..@ data       :'data.frame':	1 obs. of  1 variable:
#>   .. .. ..$ Id: int 0
#>   .. ..@ lines      :List of 1
#>   .. .. ..$ :Formal class 'Lines' [package "sp"] with 2 slots
#>   .. .. .. .. ..@ Lines:List of 14
#>   .. .. .. .. .. ..$ :Formal class 'Line' [package "sp"] with 1 slot
#>   .. .. .. .. .. .. .. ..@ coords: num [1:812, 1:2] 894793 894701 894598 894570 894516 ...
#>   .. .. .. .. .. ..$ :Formal class 'Line' [package "sp"] with 1 slot
#>   .. .. .. .. .. .. .. ..@ coords: num [1:33, 1:2] 864906 864892 864901 864899 864887 ...
#>   .. .. .. .. .. ..$ :Formal class 'Line' [package "sp"] with 1 slot
#>   .. .. .. .. .. .. .. ..@ coords: num [1:327, 1:2] 878280 878142 878076 878029 877963 ...
#>   .. .. .. .. .. ..$ :Formal class 'Line' [package "sp"] with 1 slot
#>   .. .. .. .. .. .. .. ..@ coords: num [1:674, 1:2] 878280 878281 878285 878255 878247 ...
#>   .. .. .. .. .. ..$ :Formal class 'Line' [package "sp"] with 1 slot
#>   .. .. .. .. .. .. .. ..@ coords: num [1:43, 1:2] 865525 865507 865504 865505 865492 ...
#>   .. .. .. .. .. ..$ :Formal class 'Line' [package "sp"] with 1 slot
#>   .. .. .. .. .. .. .. ..@ coords: num [1:488, 1:2] 865525 865432 865363 865266 865205 ...
#>   .. .. .. .. .. ..$ :Formal class 'Line' [package "sp"] with 1 slot
#>   .. .. .. .. .. .. .. ..@ coords: num [1:61, 1:2] 851642 851558 851522 851454 851414 ...
#>   .. .. .. .. .. ..$ :Formal class 'Line' [package "sp"] with 1 slot
#>   .. .. .. .. .. .. .. ..@ coords: num [1:543, 1:2] 851642 851635 851593 851511 851462 ...
#>   .. .. .. .. .. ..$ :Formal class 'Line' [package "sp"] with 1 slot
#>   .. .. .. .. .. .. .. ..@ coords: num [1:786, 1:2] 873422 873406 873385 873391 873405 ...
#>   .. .. .. .. .. ..$ :Formal class 'Line' [package "sp"] with 1 slot
#>   .. .. .. .. .. .. .. ..@ coords: num [1:41, 1:2] 873422 873403 873280 873263 873239 ...
#>   .. .. .. .. .. ..$ :Formal class 'Line' [package "sp"] with 1 slot
#>   .. .. .. .. .. .. .. ..@ coords: num [1:454, 1:2] 872726 872587 872560 872436 872387 ...
#>   .. .. .. .. .. ..$ :Formal class 'Line' [package "sp"] with 1 slot
#>   .. .. .. .. .. .. .. ..@ coords: num [1:484, 1:2] 872726 872710 872706 872668 872617 ...
#>   .. .. .. .. .. ..$ :Formal class 'Line' [package "sp"] with 1 slot
#>   .. .. .. .. .. .. .. ..@ coords: num [1:869, 1:2] 860552 860458 860428 860385 860210 ...
#>   .. .. .. .. .. ..$ :Formal class 'Line' [package "sp"] with 1 slot
#>   .. .. .. .. .. .. .. ..@ coords: num [1:361, 1:2] 860552 860545 860490 860488 860504 ...
#>   .. .. .. .. ..@ ID   : chr "0"
#>   .. ..@ bbox       : num [1:2, 1:2] 839774 6922087 895655 7010276
#>   .. .. ..- attr(*, "dimnames")=List of 2
#>   .. .. .. ..$ : chr [1:2] "x" "y"
#>   .. .. .. ..$ : chr [1:2] "min" "max"
#>   .. ..@ proj4string:Formal class 'CRS' [package "sp"] with 1 slot
#>   .. .. .. ..@ projargs: chr "+proj=utm +zone=5 +datum=NAD83 +units=m +no_defs +ellps=GRS80 +towgs84=0,0,0"
#>  $ lines      :List of 14
#>   ..$ : num [1:812, 1:2] 894793 894701 894598 894570 894516 ...
#>   ..$ : num [1:33, 1:2] 864906 864892 864901 864899 864887 ...
#>   ..$ : num [1:327, 1:2] 878280 878142 878076 878029 877963 ...
#>   ..$ : num [1:674, 1:2] 878280 878281 878285 878255 878247 ...
#>   ..$ : num [1:43, 1:2] 865525 865507 865504 865505 865492 ...
#>   ..$ : num [1:488, 1:2] 865525 865432 865363 865266 865205 ...
#>   ..$ : num [1:61, 1:2] 851642 851558 851522 851454 851414 ...
#>   ..$ : num [1:543, 1:2] 851642 851635 851593 851511 851462 ...
#>   ..$ : num [1:786, 1:2] 873422 873406 873385 873391 873405 ...
#>   ..$ : num [1:41, 1:2] 873422 873403 873280 873263 873239 ...
#>   ..$ : num [1:454, 1:2] 872726 872587 872560 872436 872387 ...
#>   ..$ : num [1:484, 1:2] 872726 872710 872706 872668 872617 ...
#>   ..$ : num [1:869, 1:2] 860552 860458 860428 860385 860210 ...
#>   ..$ : num [1:361, 1:2] 860552 860545 860490 860488 860504 ...
#>  $ connections: num [1:14, 1:14] NA NA 2 2 NA NA NA NA NA NA ...
#>  $ lengths    : num [1:14] 76631 1066 27718 52881 1486 ...
#>  $ names      : logi [1:14] NA NA NA NA NA NA ...
#>  $ mouth      :List of 2
#>   ..$ mouth.seg : num 1
#>   ..$ mouth.vert: num 1
#>  $ sequenced  : logi TRUE
#>  $ tolerance  : num 100
#>  $ units      : chr "m"
#>  $ lineID     :'data.frame':	14 obs. of  3 variables:
#>   ..$ rivID  : int [1:14] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ sp_line: num [1:14] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ sp_seg : int [1:14] 1 2 3 4 5 6 7 8 9 10 ...
#>  $ braided    : logi FALSE
#>  $ segroutes  :List of 14
#>   ..$ : int 1
#>   ..$ : int [1:4] 1 3 5 2
#>   ..$ : int [1:2] 1 3
#>   ..$ : int [1:2] 1 4
#>   ..$ : int [1:3] 1 3 5
#>   ..$ : int [1:3] 1 3 6
#>   ..$ : int [1:4] 1 3 6 7
#>   ..$ : int [1:4] 1 3 6 8
#>   ..$ : int [1:3] 1 4 9
#>   ..$ : int [1:3] 1 4 10
#>   ..$ : int [1:4] 1 4 10 11
#>   ..$ : int [1:4] 1 4 10 12
#>   ..$ : int [1:5] 1 4 10 11 13
#>   ..$ : int [1:5] 1 4 10 11 14
#>  $ cumuldist  :List of 14
#>   ..$ : num [1:812] 0 93.2 211.3 249.3 303 ...
#>   ..$ : num [1:33] 0 21.7 71.8 83.5 100.6 ...
#>   ..$ : num [1:327] 0 141 223 271 338 ...
#>   ..$ : num [1:674] 0 22.4 99.2 199.6 276.7 ...
#>   ..$ : num [1:43] 0 91.6 106.9 173.3 202.3 ...
#>   ..$ : num [1:488] 0 113 185 283 357 ...
#>   ..$ : num [1:61] 0 115 159 235 309 ...
#>   ..$ : num [1:543] 0 45.7 130.4 225.8 274.8 ...
#>   ..$ : num [1:786] 0 19.3 52.3 92.7 123.8 ...
#>   ..$ : num [1:41] 0 33.7 197.4 235.2 348.2 ...
#>   ..$ : num [1:454] 0 140 172 300 349 ...
#>   ..$ : num [1:484] 0 31.4 66.4 116 169.4 ...
#>   ..$ : num [1:869] 0 94.7 129.8 172.4 360.5 ...
#>   ..$ : num [1:361] 0 38.7 118.6 146.7 171.6 ...
#>  $ distlookup :List of 3
#>   ..$ middist : num [1:14, 1:14] 0 29204 0 0 27718 ...
#>   ..$ starttop: logi [1:14, 1:14] NA TRUE TRUE TRUE TRUE TRUE ...
#>   ..$ endtop  : logi [1:14, 1:14] NA FALSE FALSE FALSE FALSE FALSE ...
#>  - attr(*, "class")= chr "rivernetwork"